Aggiunto unit testing, mediante framework di testing SpringBoot e mockito. Migliorati alcuni metodi e gestiti alcuni casi particolari

+++ Model5/gestionetirocinio/src/main/java/com/gestionetirocinio/Modello/ElaboratoFinale.java
@@ -21,7 +21,7 @@ import lombok.NoArgsConstructor;
 public class ElaboratoFinale {
 
     @Id
-    private Long matricola;
+    private int matricola;
 
 /*     @OneToOne
     @JoinColumn(name = "matricola", referencedColumnName = "matricola" ) */
